requirements.txt
customtkinter==5.1.2
openai==0.27.4
openai_whisper==20230314
python-dotenv==1.0.0
requests==2.28.2
torch==2.0.0+cu118
tqdm==4.65.0
whisper==1.1.10
